feat(tree): generated TypeScript emission replaces hand-written wire types
