+++ packages/framework/src/Support/TypeScript/TypeScriptProfile.php
@@ -12,7 +12,7 @@ interface TypeScriptProfile
     /**
      * Wire packages excluded from TypeScript generation: each hand-writes
      * TypeScript that a generated.ts sibling would collide with —
-     * tree/media/api-reference augment `@lattice-php/core`'s registries
+     * media/api-reference augment `@lattice-php/core`'s registries
      * directly with no generated.ts of their own yet, and signature-example
      * (though it has no such augmentation) trips the exhaustive
      * `NodeType`/`RegisteredNodeType` check in
@@ -22,7 +22,7 @@ interface TypeScriptProfile
      *
      * @var list<string>
      */
-    public const array EMISSION_EXCLUDED = ['tree', 'media', 'api-reference', 'signature-example'];
+    public const array EMISSION_EXCLUDED = ['media', 'api-reference', 'signature-example'];
 
     /** Run a generation pass and return a summary line for the command to print. */
     public function run(): string;
