+++ packages/tree/resources/js/index.ts
@@ -1,3 +1,3 @@
 export { default as treePlugin } from "./plugin";
 export { default as Tree } from "./tree";
-export type { TreeNodeData, TreeWireProps } from "./tree";
+export type { Tree as TreeWireProps, TreeNodeData } from "./types";
