+++ packages/tree/resources/js/test-support.tsx
@@ -2,7 +2,8 @@ import type { RenderResult } from "@testing-library/react";
 import { vi } from "vitest";
 import { createRegistry, eagerComponent, type RendererComponent } from "@lattice-php/core";
 import { fakeNode, jsonResponse, renderWithRegistry } from "@lattice-php/core/test-support";
-import TreeComponent, { type TreeNodeData } from "./tree";
+import type { TreeNodeData } from "./types";
+import TreeComponent from "./tree";
 
 export const moveAction = fakeNode({
   props: { endpoint: "/lattice/actions/move", method: "post", ref: "move-ref" },
