+++ packages/tree/resources/js/tree-lazy.test.tsx
@@ -2,7 +2,8 @@ import { fireEvent, screen, waitFor } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { fakeNode, jsonResponse, renderWithRegistry } from "@lattice-php/core/test-support";
 import { testRegistry, treeNode } from "./test-support";
-import TreeComponent, { type TreeNodeData } from "./tree";
+import type { TreeNodeData } from "./types";
+import TreeComponent from "./tree";
 
 const fetchMock = vi.fn<typeof fetch>();
