+++ packages/tree/resources/js/types.ts
@@ -1,0 +1,7 @@
+import type { ComponentPropsMap } from "./generated";
+
+declare module "@lattice-php/core" {
+  interface ComponentProps extends ComponentPropsMap {}
+}
+
+export type { NodeType as TreeNodeType, Tree, TreeNodeData } from "./generated";
